refactor: change point plotting to use point struct
Make the fdf_plot_pixel function take a pointer to a point struct, so I
can pass the address of the point struct in the plot line function. This
uses less variables, and makes more sense. This will continue to be
changed so more things are using the point struct, this is mostly just a
fix.